Extract BufferedBLEParser, remove ThreadSafeMIDIBLEParser
