+++ src/MIDI_Interfaces/BLEMIDI/ArduinoBLEBackend.hpp
@@ -4,10 +4,8 @@
 
 #include "ArduinoBLE/midi.hpp"
 #include "BLEAPI.hpp"
-#include "MIDI_Interfaces/BLEMIDI/BLERingBuf.hpp"
+#include "BufferedBLEParser.hpp"
 #include "PollingMIDISender.hpp"
-#include <MIDI_Parsers/BLEMIDIParser.hpp>
-#include <MIDI_Parsers/SerialMIDI_Parser.hpp>
 
 BEGIN_CS_NAMESPACE
 
@@ -34,7 +32,7 @@ class ArduinoBLEBackend : private PollingMIDISender<ArduinoBLEBackend>,
             BLEDataView packet = data();
             if (packet.length == 0) {
                 break;
-            } else if (!ble_buffer.push(packet)) {
+            } else if (!parser.pushPacket(packet)) {
                 DEBUGREF(F("BLE packet dropped, size: ") << packet.length);
                 break;
             }
@@ -46,46 +44,11 @@ class ArduinoBLEBackend : private PollingMIDISender<ArduinoBLEBackend>,
     bool connected = false;
     /// Did the BLE Central subscribe to be notified for the MIDI characteristic?
     bool subscribed = false;
-    /// Contains incoming data to be parsed.
-    BLERingBuf<1024> ble_buffer {};
-    /// Parses the (chunked) BLE packet obtained from @ref ble_buffer.
-    BLEMIDIParser ble_parser {nullptr, 0};
-    /// Parser for MIDI data extracted from the BLE packet by @ref ble_parser.
-    SerialMIDI_Parser parser {false};
+    /// Contains incoming BLE MIDI data to be parsed.
+    BufferedBLEParser<1024> parser;
 
   public:
-    /// MIDI message variant type.
-    struct IncomingMIDIMessage {
-        MIDIReadEvent eventType = MIDIReadEvent::NO_MESSAGE;
-        union Message {
-            ChannelMessage channelmessage;
-            SysCommonMessage syscommonmessage;
-            RealTimeMessage realtimemessage;
-            SysExMessage sysexmessage;
-
-            Message() : realtimemessage(0x00) {}
-            Message(ChannelMessage msg) : channelmessage(msg) {}
-            Message(SysCommonMessage msg) : syscommonmessage(msg) {}
-            Message(RealTimeMessage msg) : realtimemessage(msg) {}
-            Message(SysExMessage msg) : sysexmessage(msg) {}
-        } message;
-        uint16_t timestamp = 0xFFFF;
-
-        IncomingMIDIMessage() = default;
-        IncomingMIDIMessage(ChannelMessage message, uint16_t timestamp)
-            : eventType(MIDIReadEvent::CHANNEL_MESSAGE), message(message),
-              timestamp(timestamp) {}
-        IncomingMIDIMessage(SysCommonMessage message, uint16_t timestamp)
-            : eventType(MIDIReadEvent::SYSCOMMON_MESSAGE), message(message),
-              timestamp(timestamp) {}
-        IncomingMIDIMessage(RealTimeMessage message, uint16_t timestamp)
-            : eventType(MIDIReadEvent::REALTIME_MESSAGE), message(message),
-              timestamp(timestamp) {}
-        IncomingMIDIMessage(SysExMessage message, uint16_t timestamp)
-            : eventType(message.isLastChunk() ? MIDIReadEvent::SYSEX_MESSAGE
-                                              : MIDIReadEvent::SYSEX_CHUNK),
-              message(message), timestamp(timestamp) {}
-    };
+    using IncomingMIDIMessage = AnyMIDIMessage;
 
     /// Retrieve and remove a single incoming MIDI message from the buffer.
     bool popMessage(IncomingMIDIMessage &incomingMessage) {
@@ -95,46 +58,8 @@ class ArduinoBLEBackend : private PollingMIDISender<ArduinoBLEBackend>,
         auto lck = Sender::acquirePacket();
         Sender::releasePacketAndNotify(lck);
         arduino_ble_midi::poll();
-        // Try reading a MIDI message from the parser
-        auto try_read = [&] {
-            MIDIReadEvent event = parser.pull(ble_parser);
-            switch (event) {
-                case MIDIReadEvent::CHANNEL_MESSAGE:
-                    incomingMessage = {parser.getChannelMessage(),
-                                       ble_parser.getTimestamp()};
-                    return true;
-                case MIDIReadEvent::SYSEX_CHUNK: // fallthrough
-                case MIDIReadEvent::SYSEX_MESSAGE:
-                    incomingMessage = {parser.getSysExMessage(),
-                                       ble_parser.getTimestamp()};
-                    return true;
-                case MIDIReadEvent::REALTIME_MESSAGE:
-                    incomingMessage = {parser.getRealTimeMessage(),
-                                       ble_parser.getTimestamp()};
-                    return true;
-                case MIDIReadEvent::SYSCOMMON_MESSAGE:
-                    incomingMessage = {parser.getSysCommonMessage(),
-                                       ble_parser.getTimestamp()};
-                    return true;
-                case MIDIReadEvent::NO_MESSAGE: return false;
-                default: break; // LCOV_EXCL_LINE
-            }
-            return false;
-        };
-        while (true) {
-            // Try reading a MIDI message from the current buffer
-            if (try_read())
-                return true; // success, incomingMessage updated
-            // Get the next chunk of the BLE packet (if available)
-            BLEDataView chunk;
-            auto popped = ble_buffer.pop(chunk);
-            if (popped == BLEDataType::None)
-                return false; // no more BLE data available
-            else if (popped == BLEDataType::Continuation)
-                ble_parser.extend(chunk.data, chunk.length); // same BLE packet
-            else if (popped == BLEDataType::Packet)
-                ble_parser = {chunk.data, chunk.length}; // new BLE packet
-        }
+        // Actually get a MIDI message from the buffer
+        return parser.popMessage(incomingMessage);
     }
 
   public:
