+++ src/MIDI_Interfaces/BLEMIDI/ESP32Backend.hpp
@@ -3,9 +3,17 @@
 #include <AH/Error/Error.hpp>
 
 #include "BLEAPI.hpp"
+#include "BufferedBLEParser.hpp"
 #include "FreeRTOSBLEMIDISender.hpp"
-#include "ThreadSafeMIDIBLEParser.hpp"
 #include "Util/ESP32Threads.hpp"
+
+#include <atomic>
+
+#ifndef ARDUINO
+#define ESP_LOGD(...) ((void)0)
+#define ESP_LOGE(...) ((void)0)
+#define ESP_LOGI(...) ((void)0)
+#endif
 
 BEGIN_CS_NAMESPACE
 
@@ -13,7 +21,8 @@ BEGIN_CS_NAMESPACE
 template <class Impl>
 class ESP32BLEBackend : private FreeRTOSBLEMIDISender<ESP32BLEBackend<Impl>>,
                         private MIDIBLEInstance {
-  private:
+  protected:
+    [[no_unique_address]] Impl impl;
     using Sender = FreeRTOSBLEMIDISender<ESP32BLEBackend>;
     friend Sender;
     void sendData(BLEDataView data) {
@@ -23,22 +32,24 @@ class ESP32BLEBackend : private FreeRTOSBLEMIDISender<ESP32BLEBackend<Impl>>,
                  chr.characteristic);
         if (chr.characteristic == 0xFFFF)
             return;
-        Impl::notify(con, chr, data);
+        impl.notify(con, chr, data);
     }
     std::atomic<BLEConnectionHandle> connection;
     std::atomic<BLECharacteristicHandle> characteristic;
 
-  private:
+  protected:
     void handleConnect(BLEConnectionHandle conn_handle) override {
         ESP_LOGD("CS-BLEMIDI", "conn=%d", conn_handle.conn);
         this->connection.store(conn_handle);
     }
-    void handleDisconnect(BLEConnectionHandle conn_handle) override {
+    void handleDisconnect(
+        [[maybe_unused]] BLEConnectionHandle conn_handle) override {
         ESP_LOGD("CS-BLEMIDI", "conn=%d", conn_handle.conn);
         this->connection.store({});
         this->characteristic.store({});
     }
-    void handleMTU(BLEConnectionHandle conn_handle, uint16_t mtu) override {
+    void handleMTU([[maybe_unused]] BLEConnectionHandle conn_handle,
+                   uint16_t mtu) override {
         ESP_LOGD("CS-BLEMIDI", "conn=%d, mtu=%d", conn_handle.conn, mtu);
         Sender::updateMTU(mtu);
     }
@@ -54,24 +65,59 @@ class ESP32BLEBackend : private FreeRTOSBLEMIDISender<ESP32BLEBackend<Impl>>,
             this->characteristic.store({});
         }
     }
-    void handleData(BLEConnectionHandle conn_handle, BLEDataGenerator &&data,
-                    BLEDataLifetime lifetime) override {
+    void handleData([[maybe_unused]] BLEConnectionHandle conn_handle,
+                    BLEDataGenerator &&data, BLEDataLifetime) override {
         ESP_LOGD("CS-BLEMIDI", "conn=%d", conn_handle.conn);
-        parser.parseNowThreadSafe(std::move(data));
+        BLEDataView packet = data();
+        if (!packet)
+            return;
+        if (!parser.pushPacket(packet)) {
+            ESP_LOGE("CS-BLEMIDI", "BLE packet dropped, size=%d",
+                     packet.length);
+            return;
+        }
+        while (BLEDataView cont = data()) {
+            if (!parser.pushPacket(cont, BLEDataType::Continuation)) {
+                ESP_LOGE("CS-BLEMIDI", "BLE chunk dropped, size=%d",
+                         cont.length);
+                return;
+            } else {
+                ESP_LOGI("CS-BLEMIDI", "added chunk, size=%d", cont.length);
+            }
+        }
     }
 
   private:
-    ThreadSafeMIDIBLEParser parser;
+    struct AtomicSize {
+#ifdef ESP32
+        constexpr static size_t alignment = 32; // default cache size
+#else
+        constexpr static size_t alignment = 64;
+#endif
+        AtomicSize(uint_fast16_t value) : value {value} {}
+        std::atomic_uint_fast16_t value;
+        uint_fast16_t load_acquire() const {
+            return value.load(std::memory_order_acquire);
+        }
+        void add_release(uint_fast16_t t) {
+            value.fetch_add(t, std::memory_order_release);
+        }
+        void sub_release(uint_fast16_t t) {
+            value.fetch_sub(t, std::memory_order_release);
+        }
+    };
+    /// Contains incoming BLE MIDI data to be parsed.
+    BufferedBLEParser<4096, AtomicSize> parser;
 
   public:
-    using IncomingMIDIMessage = ThreadSafeMIDIBLEParser::IncomingMIDIMessage;
+    using IncomingMIDIMessage = AnyMIDIMessage;
     bool popMessage(IncomingMIDIMessage &incomingMessage) {
         return parser.popMessage(incomingMessage);
     }
 
   public:
     void begin(BLESettings ble_settings) {
-        Impl::init(*this, ble_settings);
+        impl.init(*this, ble_settings);
         // Need larger stack than default, pin to non-Arduino core
         ScopedThreadConfig sc {4096, 3, true, "CS-BLEMIDI", 0};
         Sender::begin();
